Added a reviewscontroller to display all the reviews.
Added a Restaurant Reviews object to hold all the review data,

Added a index view to show all the views to the users.

Added a edit view to allow users to edit the data in the views.

modified the _layout view to allow us to display a feature div after tha main body and also to add a tab to redirect to the reviews/index page.

Learned about html helpers and how they can be used to display our objects securely on a view. Also the TryUpdateModel method which can be used to apply all the user edits to the model. I will return false if the update failed.