Added 2 objects for Restaurant and RestaurantReviews.
Added one OdeToFoodDB class which derives off of DbContext that will hold a dbSet of Restaurant and another DbSet of RestaurantReviews.

The EF made a db in the myLocal SQL server (localdb)\mssqllocaldb with the tables to match my code. Even the foreign keys have been created properly. It's Amazing! :D

Removed most of the code from the home page view ans well as the review controller. That code needs to be rewrtten to match the requirements now using the EF and SQL db.