Added references to jquery ui as well as knockout
added the jquery UI to the budleconfig.

added a javascrip file called otf..js that has custom java script to handle the auto complete suggestions in the search box,

it also has the code to populate the bottom of the screen without the whole screen in ajaxFormSubmit.,

it also has a function to call the submit form when a user selects any of the autocomplete suggestions.