Created an entire unit test for the home controller index action.
We made a interface to choose who we are connecting to.

We adde a fake db with fake data.

added a fake context to pass httpcontext off to the controller to check the IsAjax method.

Lots of stuff going on over here, Things I need to look at properly again to understand and apply.