Created Migrations for deployment,
Learned how to deploy the web app to IIS running on my local machine. Using CMD to push it to IIS.

Tried to put it up on Windows Azure but it didnt work as planned. I have to read a bit more to find out what goes where and how to setupt the Azure account.

I feel a little lost. Maybe I should try a project before starting on a new course?? :/